refactor(server): extract the language tool into src/server/handlers/language-tools.ts
list_supported_languages, handleListSupportedLanguages and the
AvailableLanguage entry type move off the class verbatim (`this.` ->
`ctx.`). The class keeps a one-line private delegate as the test seam
(removed in PR 6).
